retriever/requirements.txt
sodapy
psycopg2-binary